requirements.txt
langchain 
requests 
openai
langchain_community
langchain-openai
python-dotenv
langchain-openai
chromadb